# src/config/openapi.yaml
openapi: 3.0.0
info:
  title: Dating App API
  description: API documentation for the Dating App backend (Express.js, Sequelize)
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local server
  - url: /
    description: Relative path for deployed environments
security:
  - bearerAuth: []
paths:
...
  /api/users/discover:
    get:
      summary: Get available users to discover
      tags: [Users Discovery]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Users retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UsersListResponseData"
  /api/users/recommended:
    get:
      summary: Get recommended users
      tags: [Users Discovery]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recommended users retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UsersListResponseData"
  /api/users/{userId}:
    get:
      summary: Get user details by ID
      tags: [Users Discovery]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserFormatted"
  /api/users/me:
    get:
      summary: Get current authenticated user profile
      tags: [Users Discovery]
      security:
        - bearerAuth: []

      responses:
        "200":
          description: Current user profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserFormatted"

        "401":
          description: Unauthorized

  # ==========================================
  # USER PREFERENCES ROUTES
  # ==========================================
  /api/users/preferences:
    get:
      summary: Get current user preferences
      tags: [User Preferences]
      responses:
        "200":
          description: Preferences retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserPreferences"
    post:
      summary: Create user preferences
      tags: [User Preferences]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                target_gender:
                  type: string
                  enum: [male, female, other]
                min_age:
                  type: integer
                max_age:
                  type: integer
                max_distance_km:
                  type: integer
                anonymous_interests:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Preferences created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserPreferences"
    put:
      summary: Update user preferences
      tags: [User Preferences]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                target_gender:
                  type: string
                  enum: [male, female, other]
                min_age:
                  type: integer
                max_age:
                  type: integer
                max_distance_km:
                  type: integer
                anonymous_interests:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserPreferences"

  # ==========================================
  # USER PHOTOS ROUTES
  # ==========================================
  /api/users/photos:
    get:
      summary: Get all user photos
      tags: [User Photos]
      responses:
        "200":
          description: Photos retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Photo"
    post:
      summary: Upload/Create a new user photo
      tags: [User Photos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_url]
              properties:
                image_url:
                  type: string
                is_primary:
                  type: boolean
                display_order:
                  type: integer
      responses:
        "201":
          description: Photo created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Photo"
  /api/users/photos/{photoId}:
    put:
      summary: Update user photo details
      tags: [User Photos]
      parameters:
        - in: path
          name: photoId
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                image_url:
                  type: string
                is_primary:
                  type: boolean
                display_order:
                  type: integer
      responses:
        "200":
          description: Photo updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Photo"
    delete:
      summary: Delete a user photo
      tags: [User Photos]
      parameters:
        - in: path
          name: photoId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Photo deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/users/photos/{photoId}/primary:
    put:
      summary: Set a specific photo as primary (avatar)
      tags: [User Photos]
      parameters:
        - in: path
          name: photoId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Photo set as primary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Photo"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
          nullable: true
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 50
        totalPages:
          type: integer
          example: 5

    # --- BASE MODELS ---
    UserBase:
      type: object
      properties:
        user_id:
          type: integer
        email:
          type: string
        full_name:
          type: string
        birth_date:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        bio:
          type: string
        default_mode:
          type: string
        created_at:
          type: string
          format: date-time
    UserFormatted:
      allOf:
        - $ref: "#/components/schemas/UserBase"
        - type: object
          properties:
            preferences:
              $ref: "#/components/schemas/UserPreferences"
            photos:
              type: array
              items:
                $ref: "#/components/schemas/Photo"
    Photo:
      type: object
      properties:
        photo_id:
          type: integer
        user_id:
          type: integer
        image_url:
          type: string
        is_primary:
          type: boolean
        display_order:
          type: integer
          nullable: true
    UserPreferences:
      type: object
      properties:
        preference_id:
          type: integer
        target_gender:
          type: string
        min_age:
          type: integer
        max_age:
          type: integer
        max_distance_km:
          type: integer
        anonymous_interests:
          type: array
          items:
            type: string
...
    UsersListResponseData:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserFormatted"
        pagination:
          $ref: "#/components/schemas/PaginationInfo"
...